### 2016-09-12 #60 @whedon
[REVIEW]: hebbRNN: A Reward-Modulated Hebbian Learning Rule for Recurrent Neural Networks
**Submitting author:** JonathanAMichaels (<a href="http://orcid.org/0000-0002-5179-3181">Jonathan A Michaels</a>)
**Repository:** <a href="https://github.com/JonathanAMichaels/hebbRNN" target ="_blank">https://github.com/JonathanAMichaels/hebbRNN</a>
**Version:** v1.2

**Reviewers and authors**: Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue
in the <a href="https://github.com/JonathanAMichaels/hebbRNN" target="_blank">target repository</a> and link to those issues (especially acceptance-blockers)
in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice
versa will create corresponding breadcrumb trails in the link target.)
