### 2016-09-12 39d2a50
Updating paper references to fix JOSS paper compilation
