### 2016-09-22 #60 @cMadan
@JonathanAMichaels - sorry for the delay in reviewing your submission.

I looked at the documentation, tried the examples, and checked over everything--your submission impressively follows the JOSS guidelines as-submitted!

I have only two minor suggestions that need to be addressed:
(1) paper.bib, references are missing DOIs.
(2) README.md, typo: "dependecies"

Pending these very minor changes, I am happy to recommend the submission be accepted at JOSS.
