### 2016-09-23 #60 @JonathanAMichaels
Thank you for the very positive review!

I have corrected both errors with the exception of one reference that does not have a DOI.
