### 2016-09-23 #60 @cMadan
Looks good to me!

@arfon - I am happy to recommend this submission be accepted at JOSS.
