### 2016-09-23 #60 @arfon
> @arfon - I am happy to recommend this submission be accepted at JOSS.

@cMadan thanks for turning this around quickly.

@JonathanAMichaels at this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
