### 2016-09-23 #60 @arfon
Thanks again for the review @cMadan!

@JonathanAMichaels your paper is now accepted into JOSS. Your paper DOI is http://dx.doi.org/10.21105/joss.00060 🎉 🚀 💥
